Conflicts:
	src/cc1/Makefile
	src/cc1/cc1.c
	src/cc1/tokconv.c
	src/cpp2/Makefile
	src/cpp2/main.c
	src/ucc/Makefile
	src/ucc/ucc.c
	src/util/Makefile
	src/util/platform.c
	src/util/platform.h

+++ src/util/alloc.c
@@ -19,7 +19,7 @@ void *ucalloc(size_t n, size_t sz)
 {
 	void *p = calloc(n, sz);
 	if(!p)
-		ICE("calloc(%ld, %ld):", n, sz);
+		ICE("calloc(%ld, %ld):", (long)n, (long)sz);
 	return p;
 }
